fix: graphic cord
