+++ lib/presenter/home_page/repository/coin_repository.dart
@@ -1,7 +1,9 @@
+import 'coord_graphic.dart';
+
 import '../model/coin_model.dart';
 
 class CoinRepository {
-  List<CoinModel> getCoins() {
+  List<CoinModel> coins() {
     List<CoinModel> coins = [
       CoinModel(
         nameCoin: 'Bitcoin',
@@ -10,108 +12,7 @@ class CoinRepository {
         coinVariation: 8.00,
         priceCurrent: (1031.21),
         coinQuantity: (0.51),
-        coord: [
-          (13.21),
-          (0.21),
-          (20.21),
-          (1.63),
-          (4.00),
-          (5.21),
-          (6.63),
-          (16.63),
-          (7.21),
-          (8.21),
-          (9.00),
-          (10.21),
-          (3.21),
-          (11.63),
-          (14.00),
-          (15.21),
-          (17.21),
-          (12.21),
-          (19.00),
-          (20.21),
-          (21.63),
-          (25.21),
-          (22.21),
-          (23.21),
-          (24.00),
-          (18.21),
-          (38.21),
-          (26.63),
-          (27.21),
-          (28.21),
-          (29.00),
-          (30.21),
-          (32.21),
-          (33.21),
-          (34.00),
-          (35.21),
-          (36.63),
-          (37.21),
-          (31.63),
-          (39.00),
-          (40.21),
-          (41.63),
-          (42.21),
-          (43.21),
-          (44.00),
-          (45.21),
-          (46.63),
-          (47.21),
-          (48.21),
-          (49.00),
-          (50.21),
-          (51.63),
-          (80.63),
-          (53.21),
-          (64.00),
-          (54.00),
-          (56.63),
-          (57.21),
-          (58.21),
-          (59.00),
-          (60.21),
-          (61.63),
-          (62.21),
-          (63.21),
-          (52.21),
-          (66.63),
-          (67.21),
-          (68.21),
-          (69.00),
-          (70.21),
-          (87.21),
-          (55.21),
-          (81.21),
-          (82.21),
-          (83.00),
-          (84.21),
-          (85.63),
-          (86.21),
-          (80.63),
-          (87.21),
-          (70.21),
-          (84.21),
-          (65.21),
-          (81.21),
-          (82.21),
-          (88.00),
-          (83.00),
-          (85.63),
-          (83.00),
-          (88.00),
-          (70.21),
-          (80.63),
-          (82.21),
-          (86.21),
-          (81.21),
-          (86.21),
-          (85.63),
-          (84.21),
-          (87.21),
-          (88.00),
-        ],
+        coord: coord,
       ),
       CoinModel(
         nameCoin: 'Ethereum',
@@ -120,108 +21,7 @@ class CoinRepository {
         coinVariation: -222.25,
         priceCurrent: (799.96),
         coinQuantity: (0.51),
-        coord: [
-          (13.21),
-          (0.21),
-          (20.21),
-          (1.63),
-          (4.00),
-          (5.21),
-          (6.63),
-          (16.63),
-          (7.21),
-          (8.21),
-          (9.00),
-          (10.21),
-          (3.21),
-          (11.63),
-          (14.00),
-          (15.21),
-          (17.21),
-          (12.21),
-          (19.00),
-          (20.21),
-          (21.63),
-          (25.21),
-          (22.21),
-          (23.21),
-          (24.00),
-          (18.21),
-          (38.21),
-          (26.63),
-          (27.21),
-          (28.21),
-          (29.00),
-          (30.21),
-          (32.21),
-          (33.21),
-          (34.00),
-          (35.21),
-          (36.63),
-          (37.21),
-          (31.63),
-          (39.00),
-          (40.21),
-          (41.63),
-          (42.21),
-          (43.21),
-          (44.00),
-          (45.21),
-          (46.63),
-          (47.21),
-          (48.21),
-          (49.00),
-          (50.21),
-          (51.63),
-          (80.63),
-          (53.21),
-          (64.00),
-          (54.00),
-          (56.63),
-          (57.21),
-          (58.21),
-          (59.00),
-          (60.21),
-          (61.63),
-          (62.21),
-          (63.21),
-          (52.21),
-          (66.63),
-          (67.21),
-          (68.21),
-          (69.00),
-          (70.21),
-          (87.21),
-          (55.21),
-          (81.21),
-          (82.21),
-          (83.00),
-          (84.21),
-          (85.63),
-          (86.21),
-          (80.63),
-          (87.21),
-          (70.21),
-          (84.21),
-          (65.21),
-          (81.21),
-          (82.21),
-          (88.00),
-          (83.00),
-          (85.63),
-          (83.00),
-          (88.00),
-          (70.21),
-          (80.63),
-          (82.21),
-          (86.21),
-          (81.21),
-          (86.21),
-          (85.63),
-          (84.21),
-          (87.21),
-          (88.00),
-        ],
+        coord: coord,
       ),
       CoinModel(
         nameCoin: 'Litecoin',
@@ -230,108 +30,7 @@ class CoinRepository {
         coinVariation: 82.9,
         priceCurrent: (245.01),
         coinQuantity: (0.51),
-        coord: [
-          (13.21),
-          (0.21),
-          (20.21),
-          (1.63),
-          (4.00),
-          (5.21),
-          (6.63),
-          (16.63),
-          (7.21),
-          (8.21),
-          (9.00),
-          (10.21),
-          (3.21),
-          (11.63),
-          (14.00),
-          (15.21),
-          (17.21),
-          (12.21),
-          (19.00),
-          (20.21),
-          (21.63),
-          (25.21),
-          (22.21),
-          (23.21),
-          (24.00),
-          (18.21),
-          (38.21),
-          (26.63),
-          (27.21),
-          (28.21),
-          (29.00),
-          (30.21),
-          (32.21),
-          (33.21),
-          (34.00),
-          (35.21),
-          (36.63),
-          (37.21),
-          (31.63),
-          (39.00),
-          (40.21),
-          (41.63),
-          (42.21),
-          (43.21),
-          (44.00),
-          (45.21),
-          (46.63),
-          (47.21),
-          (48.21),
-          (49.00),
-          (50.21),
-          (51.63),
-          (80.63),
-          (53.21),
-          (64.00),
-          (54.00),
-          (56.63),
-          (57.21),
-          (58.21),
-          (59.00),
-          (60.21),
-          (61.63),
-          (62.21),
-          (63.21),
-          (52.21),
-          (66.63),
-          (67.21),
-          (68.21),
-          (69.00),
-          (70.21),
-          (87.21),
-          (55.21),
-          (81.21),
-          (82.21),
-          (83.00),
-          (84.21),
-          (85.63),
-          (86.21),
-          (80.63),
-          (87.21),
-          (70.21),
-          (84.21),
-          (65.21),
-          (81.21),
-          (82.21),
-          (88.00),
-          (83.00),
-          (85.63),
-          (83.00),
-          (88.00),
-          (70.21),
-          (80.63),
-          (82.21),
-          (86.21),
-          (81.21),
-          (86.21),
-          (85.63),
-          (84.21),
-          (87.21),
-          (88.00),
-        ],
+        coord: coord,
       ),
     ];
     return coins;
