+++ lib/presenter/home_page/repository/coord_graphic.dart
@@ -1,0 +1,102 @@
+List<double> coord = [
+  (10.14),
+  (5.17),
+  (13.14),
+  (17.11),
+  (3.10),
+  (15.33),
+  (10.43),
+  (16.22),
+  (7.21),
+  (17.42),
+  (20.32),
+  (12.10),
+  (3.11),
+  (11.12),
+  (1.20),
+  (15.17),
+  (17.21),
+  (12.23),
+  (16.32),
+  (22.12),
+  (21.42),
+  (43.13),
+  (22.21),
+  (23.21),
+  (24.34),
+  (18.21),
+  (38.21),
+  (26.63),
+  (27.21),
+  (28.21),
+  (29.22),
+  (30.21),
+  (32.21),
+  (33.21),
+  (34.24),
+  (35.21),
+  (36.12),
+  (37.32),
+  (31.11),
+  (39.43),
+  (40.21),
+  (41.12),
+  (42.21),
+  (43.21),
+  (44.33),
+  (42.21),
+  (33.22),
+  (11.21),
+  (33.21),
+  (21.00),
+  (42.21),
+  (36.32),
+  (43.42),
+  (53.21),
+  (32.32),
+  (22.12),
+  (14.63),
+  (21.21),
+  (58.21),
+  (59.00),
+  (22.21),
+  (31.12),
+  (22.21),
+  (44.21),
+  (46.21),
+  (33.14),
+  (16.21),
+  (34.21),
+  (54.00),
+  (12.21),
+  (23.21),
+  (35.21),
+  (64.21),
+  (21.21),
+  (54.00),
+  (23.21),
+  (44.15),
+  (12.21),
+  (33.32),
+  (22.21),
+  (23.21),
+  (32.21),
+  (42.21),
+  (16.21),
+  (23.21),
+  (45.11),
+  (19.22),
+  (22.22),
+  (27.40),
+  (49.42),
+  (34.21),
+  (55.32),
+  (13.21),
+  (12.21),
+  (33.21),
+  (41.21),
+  (23.11),
+  (45.21),
+  (54.21),
+  (21.00),
+];
