+++ lib/presenter/home_page/widgets/coin_list.dart
@@ -1,10 +1,9 @@
-import 'package:crypto_screen_card_1/presenter/home_page/widgets/coin_widgets/coin_icon.dart';
-import 'package:crypto_screen_card_1/presenter/home_page/widgets/coin_widgets/coin_names.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 import '../model/coin_model.dart';
-import '../provider/provider.dart';
+import 'coin_widgets/coin_icon.dart';
+import 'coin_widgets/coin_names.dart';
 
 class CoinsList extends HookConsumerWidget {
   final CoinModel model;
@@ -14,8 +13,6 @@ class CoinsList extends HookConsumerWidget {
   });
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    final walletController = ref.watch(walletControllerProvider);
-
     return Padding(
       padding: const EdgeInsets.all(10),
       child: Row(
