+++ lib/presenter/home_page/widgets/coin_widgets/coin_names.dart
@@ -1,8 +1,8 @@
-import 'package:crypto_screen_card_1/presenter/home_page/widgets/coin_widgets/coin_initials.dart';
 import 'package:flutter/material.dart';
-import '../../model/coin_model.dart';
 
+import '../../model/coin_model.dart';
 import 'coin_price_current.dart';
+import 'coin_single_value.dart';
 
 class CoinNames extends StatelessWidget {
   const CoinNames({
@@ -41,7 +41,7 @@ class CoinNames extends StatelessWidget {
                   color: Colors.grey,
                 ),
               ),
-              CoinPercentage(coin: model),
+              SingleCoinValue(coin: model),
             ],
           ),
         ],
